Switch BrowsingContextRequest to full parameter DTOs

+++ src/main/java/wd4j/impl/webdriver/command/request/parameters/browsingContext/TraverseHistoryParameters.java
@@ -3,11 +3,14 @@ package wd4j.impl.webdriver.command.request.parameters.browsingContext;
 import wd4j.impl.webdriver.type.browsingContext.BrowsingContext;
 import wd4j.impl.websocket.Command;
 
+/**
+ * The browsingContext.traverseHistory command traverses the history of a given navigable by a delta.
+ */
 public class TraverseHistoryParameters implements Command.Params {
     private final BrowsingContext browsingContext;
-    private final int delta;
+    private final long delta;
 
-    public TraverseHistoryParameters(BrowsingContext browsingContext, int delta) {
+    public TraverseHistoryParameters(BrowsingContext browsingContext, long delta) {
         this.browsingContext = browsingContext;
         this.delta = delta;
     }
@@ -16,7 +19,7 @@ public class TraverseHistoryParameters implements Command.Params {
         return browsingContext;
     }
 
-    public int getDelta() {
+    public long getDelta() {
         return delta;
     }
 }
